Update the type-extensions
